+++ compat_after_go1.8.go
@@ -1,3 +1,5 @@
+// +build go1.8
+
 // Copyright 2013-2017 Aerospike, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,17 +16,10 @@
 
 package aerospike
 
-// BufferEx is a specialized buffer interface for aerospike client.
-type BufferEx interface {
-	WriteInt64(num int64) (int, error)
-	WriteUint64(num uint64) (int, error)
-	WriteInt32(num int32) (int, error)
-	WriteUint32(num uint32) (int, error)
-	WriteInt16(num int16) (int, error)
-	WriteUint16(num uint16) (int, error)
-	WriteFloat32(float float32) (int, error)
-	WriteFloat64(float float64) (int, error)
-	WriteByte(b byte) error
-	WriteString(s string) (int, error)
-	Write(b []byte) (int, error)
+import (
+	"crypto/tls"
+)
+
+func cloneTlsConfig(c *tls.Config) *tls.Config {
+	return c.Clone()
 }