+++ connection.go
@@ -93,7 +93,7 @@ func NewConnection(address string, timeout time.Duration) (*Connection, error) {
 // If the connection is not established in the specified timeout,
 // an error will be returned
 func NewSecureConnection(policy *ClientPolicy, host *Host) (*Connection, error) {
-	address := host.Name + ":" + strconv.Itoa(host.Port)
+	address := net.JoinHostPort(host.Name, strconv.Itoa(host.Port))
 	conn, err := NewConnection(address, policy.Timeout)
 	if err != nil {
 		return nil, err
@@ -103,11 +103,11 @@ func NewSecureConnection(policy *ClientPolicy, host *Host) (*Connection, error) 
 		return conn, nil
 	}
 
-	// To be on the safe side
-	tlsConfig := *policy.TlsConfig
+	// Use version dependent clone function to clone the config
+	tlsConfig := cloneTlsConfig(policy.TlsConfig)
 	tlsConfig.ServerName = host.TLSName
 
-	sconn := tls.Client(conn.conn, &tlsConfig)
+	sconn := tls.Client(conn.conn, tlsConfig)
 	if err := sconn.Handshake(); err != nil {
 		sconn.Close()
 		return nil, err